fix height in rv32_div_rem_tracegen

+++ extensions/rv32im/circuit/cuda/src/divrem.cu
@@ -234,7 +234,7 @@ struct Rv32DivRemRecord {
 
 __global__ void rv32_div_rem_tracegen(
     Fp *d_trace,
-    uint32_t height,
+    size_t height,
     DeviceBufferConstView<Rv32DivRemRecord> d_records,
     uint32_t *d_range_checker_ptr,
     uint32_t range_checker_bits,
